# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Coffee Website API Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/products/category/{category_id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "دریافت محصولات بر اساس دسته‌بندی",
                "operationId": "770934ff9827007f1b5cbc1c7d3f74ff",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "path",
                        "description": "آیدی دسته‌بندی محصول",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 3
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "لیست محصولات این دسته‌بندی",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 5
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "مانیتور سامسونگ"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "مانیتور ۲۷ اینچ 4K"
                                            },
                                            "price": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 4200000
                                            },
                                            "discount": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 15
                                            },
                                            "image": {
                                                "type": "string",
                                                "example": "monitor.jpg"
                                            },
                                            "category_id": {
                                                "type": "integer",
                                                "example": 3
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "دریافت محصول بر اساس آیدی",
                "operationId": "1948ab0c6b14a722ad6d1665cd38e1f1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "آیدی محصول",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "اطلاعات محصول",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 10
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "هدفون بی‌سیم"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "هدفون نویز کنسلینگ حرفه‌ای"
                                            },
                                            "price": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 1490000
                                            },
                                            "discount": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 5
                                            },
                                            "image": {
                                                "type": "string",
                                                "example": "headphone.jpg"
                                            },
                                            "category_id": {
                                                "type": "integer",
                                                "example": 4
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "محصول پیدا نشد"
                    }
                }
            }
        },
...
    },
...
}